+++ covetools/CHANGES.md
@@ -2,7 +2,8 @@
 
 ## 1.1.1
  * **(Breaking)** Make AssignmentAssetManager's annotations and listener into outer classes for cleaner IDE importing. 
-
+ * **(Breaking)** Add `iconDropped` method to LiveWallpaperListener.
+ 
 ## 1.1.0
  * Fixed AssignmentAssetManager to match documentation about not unloading individual assets until all containers are unloaded.
  * Add classes for easier live wallpaper development. Added `covetools-android` repository.