+++ flexbatch/CHANGES.md
@@ -4,6 +4,7 @@
  * Add `getRenderContext()` so additional parameters can be changed manually.
  * Allow a flush of the previous content of the batch with `repeatPreviousFlush()`. This saves a lot of CPU when redrawing content that is known to be unchanged since the last flush.
  * Add capacity getter methods.
+ * **(Breaking)** Renamed `flexBatch.setBlendFunction(int, int, int, int)` to `flexBatch.setBlendFunctionSeparate(int, int, int, int)` to match new method name in SpriteBatch.
 
 ## 1.1.0
  * Bugfix FlexBatch applying render context changes to the GL state outside of begin/end.